Allow deploying different templates of hubs
Offering multiple 'hub templates' for deployment makes life
much easier for us & our users. We can represent each type with
its own deployment chart, and have them depend on each other
to form a hierarchy of them.

Currently, it is only possible to hade a one-level hierarchy,
since `base-hub` is hardcoded in the code. This is config that
must be separated.

We add the 'ephemeral-hub' template, and make the ephemeral hub
use it.