Add paleohack2021 hackathon hub
We want to pass through cloud costs to our users.
Putting them in their own cloud project is the cleanest
and most transparent way to do this. I've created a project
called 'hackathon-2i2c-project-alpha' for it, and set up
the hub there. This PR adds the secret service account and
the hubs.yaml config for it.

I've also manually created the NFS directory!