Remove 'ephemeral hub' type
This was supposed to be a 'binder-like' hub, with
the following config:

1. No authentication - uses tmpauthenticator by default to let
   anyone in
2. No persistent storage!
3. More aggressive culling of users

The goal was to provide 'demonstration' hubs for workshops,
etc. However, there has been no current *demand* for this kinda
hub, and the workshops we've supported definitely wanted authentication.
The amount of customization needed is also fairly small, and can
be done on a per-hub basies without much effort.

So, let's just remove this hub type and simplfiy our code.