Generate carbonplan kops file with jsonnet
Each kops cluster we want to deploy will have some things in
common, but many differences. 2i2c engineers should be able to
create and manage these without a lot of cognitive overhead.

1. All clusters should have some basic opinionated defaults
2. But everything should be customizable as necessary
3. kops configuration is already declerative, and we should
   *not* add another layer on top.

jsonnet helps with this. It's a powerful declarative language
generating JSON, with conditionals, prototype inheritance,
deep merging, maps, etc.

In this PR, we create two base objects that can be customized -
a `cluster` object that generates a kops Cluster object,
and a `instancegroup` object that generates a kops InstanceGroup
object. We use these to generate our kops config for
carbonplan.jsonnet. jsonnet is only used for some abstraction
and deep merging - there isn't an extra level of abstraction
here.