Set proxy.https.hosts by default too
In some hubs, we currently just use a loadbalancer + autohttps
to get traffic into the cluster. This is simpler than getting
nginx-ingress + certmanager setup, although not entirely
sure if that's the right thing to do long term.

By putting the domains in proxy.https.hosts as well, we can
decide to use or not use autohttps on a per-hub basis, without
having to repeat the domains in multiple places. Staging and
prod hubs that have the exact same config but differ in
domains can be easily constructed thus.