Don't overwrite user's KUBECONFIG
Without this, your local authentication with kubectl
is reset to that of the serviceaccount used to deploy the
cluster. If you try and switch cluster, it'll fail with
strange errors.

Now we properly set the env vars while authenticated so
it does not leak out.

Found because I turned on flake8 warning for unused
local variables!