Push preliminary AWS deployment documentation
This is the first iteration in the creation of a how-to guide describing
how you can manually set up a AWS kops-based cluster and use our
deployer to manually deploy a new hub on top of it.

This is yet a very preliminary draft but with ppor styling but the
instructions should be enough detailed so others can reproduce the
process without major issues.

Will submit this effort as a draft so others can actively contribute not
only in the styling but also in the semantics (you will quickly notice
this is not my mother language, so open to critics so I can improve
myself ;-)