Separate NFS PV and PVC
We want to use Azure Files for persistent storage on
Azure, and use the native mounting available via SMB
rather than via NFS. We need a PVC that can be bound to
an existing PV in the cluster (created via terraform).

Separating the PV and PVC this way helps accomplish
that cleanly. In the future, we mmight be able to use this
for EFS too