Create AzureFile PVC / PV with Helm
One PV can't be bound to more than one PV, so we need one
PV per hub anyway. So we use terraform to create the secret,
but not the PVs.

Anytime we have a PVC per hub, we'll need a PV too. The
split of NFS PVC and PV is now unbound.