Use eksctl schema for spot instances too
For spot instances, `instanceType` can not be set on the
nodePool object, or eksctl complains. Before this commit,
we were still just specifying 'instanceType', but magically
converting it to the appropriate incantation for spot instances.
This leads to confusion, and missing features (for example,
the ability to specify multiple instanceTypes).

This commit switches us to using just plain eksctl schema,
at the cost of a slightly more verbose config