Auth AWS EKS cluster with encrypted AWS credentials
I have added a new auth_aws function following the existing pattern for
the general auth context manager. The main idea was to get a temporary
kubeconfig file with the update-kubeconfig AWS CLI command [1] so we can
successfully connect with the EKS cluster. The command looks for the IAM
entity in your default AWS CLI to properly authenticate. In this case,
we are going to provide some 2i2c-deployer AWS credentials in an sops
encrypted json file and set up the Access Key ID and Secret Access Key
as environment variables.

I have also added some new values to the schema with the aim to replace
the current auth_kubeconfig with a future (more general) auth_aws
function working for kops-based clusters as well.