Add encrypted json file with credentials
The auth_aws function will retrieve the credentials from a new
"deployer" user living under the 2i2c group (in AWS carbonplan land).
That user also needs admin access to the EKS cluster to get the
kubeconfig.
Details about how to create the user, get credentials and add it to the
EKS cluster live in [1].

Btw, I have modified the auth_aws function a little bit so it can find
the needed information in the "native" credentials json file produced
by the awscli tool. The files is encrypted by sops.

Finally, I added an enum to the schema because I believe we would just
have eks and kops options in the future.

[1] https://github.com/2i2c-org/pilot-hubs/issues/632#issuecomment-911649179