Set up the KUBECONFIG env variable just before the update-kubeconfig cmd
Trying to minimize any possibility of a race condition because the sops
file taking too much time to be decrypted, we decided to set up the
KUBECONFIG environment variable just before creating the temporary
kubeconfig file with the aws eks update-kubeconfig command.

Additionally, we update the docstring with some more information.