Add a new required stateStore property if kops is enabled
We currently manage EKS and kops clusters on AWS land. With this commit
we are adding support and validation (by jsonschema) for the kops-based
cluster that requires the state store to properly get the state of the
cluster and succeed in the next step at the time to get export the
kubeconfig.

Notice I am using the if and then keywords [1] to require the stateStore
property only if the kops value is configured.

The encrypted key file referenced in the condfig file will be provided
in a later commit.

[1] https://json-schema.org/understanding-json-schema/reference/conditionals.html#if-then-else