Add the encrypted file containing deployer credentials
We identified a 2i2c-engineers groups in Farallon AWS and we have
created a deployer user under that group. We used common awscli commands
to perform this task and retrieve the credentials. More details in [1].

Finally, we encrypted the file with sops accordingly with the current
established workflows to manage secret files.

[1] https://github.com/2i2c-org/pilot-hubs/issues/381#issuecomment-917170711