Add logic in the auth_aws function to also export kops kubeconfig
Previously, we have added an auth_aws function to retrieve the
kubeconfig and connect to EKS-based cluster. With this commit, we are
adding a new subprocess "brach" to conditionally `kops export` the
kubeconfig (if the kops option is configured) and get access to
kops-based clusters, such as the Farallon one.