Redirect test output when we are running on the CI
We were redirecting the test output to prevent the leakage of secrets
when running the deployer on the CI. At the same time, the redirection
was preventing us to get details about the failure at the time to run
the deployer locally. So, the idea of this PR is just redirecting to
dev/null only if we are running the deployer on the CI and get the full
output when running locally.