Set prometheus server's deployment strategy to recreate
During deployment of https://github.com/2i2c-org/pilot-hubs/pull/720,
sometimes CI would fail because the prometheus pod would be
stuck in 'ContainerCreating', as the old pod was holding on to
the persistent disk the new pod needs to start. This was temporarily
fixed by deleting the prometheus pod, but this tells kubernetes
to delete the old pod properly first before starting the new one.

Ref https://github.com/2i2c-org/pilot-hubs/pull/720#issuecomment-930946399