Reorder stuff to separate reference from how-to sections
Thanks to Chris' feedback I was able to reorganize the documentation I
have written in a minor (placeholder) reference section for our CI/CD
system and a how-to guide to set up the CI/CD stuff on AWS clusters.

In addition, I have tried to fix some identation problem so I get the
expected numbering in the subsections, I tried to be more specific in
the referenced (hiperlinked) sections and some other minor fixes such as
new notes or removing superfluos admonitions.