Reduce guarantee memory to avoid going beyond the allocatable value
Each node has some allocatable resources [1]
If you try to schedule a pod with guarantee requirements above the
allocatable values, it will fail to be spawned.
With the "very large" option, we are requesting 110G at minimum and that
is dangerously close the the theoretical allocatable memory for the
n1-standard-32 node. Then, let's give that guarantee value some
breath ;-)

[1] https://cloud.google.com/kubernetes-engine/docs/concepts/cluster-architecture#memory_cpu