Block access to howard cloudbank hub
The instructor wants the students to not be able to log in,
while preserving the ability to use the admin panel to access
students' work. By moving the hub to a randomly generated
domain name that is kept secret, the instructor can log in but
students can not.

Longer term, we need better user profile support in JupyterHub
for this. Perhaps the work in JupyterHub 2.0 with groups
will help.

Ref https://2i2c.freshdesk.com/a/tickets/61