Bump postgres resource limit for utexas
Increase postgresql limits by a bit - it is possible that
the kernel deaths are caused by the postgres container hitting
its memory limit and dying. The graph in
https://github.com/2i2c-org/infrastructure/issues/1021#issuecomment-1048540620
makes it look like it isn't hitting the limit, but perhaps node
memory pressure is making it get killed just past the request
(which was 64MB) - so the bump might still help.