Additional safeguards against accidental cluster deletion
I ran `terraform workspace new` to create a new workspace,
and somehow assumed it also switched me to the new workspace -
which it *does not*. So just running `terraform apply` almost
deleted a bunch of running clusters.

This lifecycle hook will *prevent* any of these from being
deleted ever. To delete these resources (as part of a decomission),
you'll need to edit the files to set it to false. This is a manual
step, but very appropriate for irreversible data loss.