Switch to using pd-balanced for all user & dask nodes
We were using standard disk to save costs, but that brings with
it much slower node startup time, as images being pulled take time.
pd-balanced is a newer alternative to pure SSD disks that is
not as expensive, but provides much better performance than pd-standard.

I think the extra cost is worth the performance on all these
cases.