Disable user scheduler almost everywhere
After https://github.com/2i2c-org/infrastructure/pull/1026,
all our GCP hubs use the 'Optimize Utilization' scheduler.
It functions pretty similar to the z2jh custom user scheduler,
and is generally faster & has less overhead. So we want to use
that wherever possible.

In other cloud providers, most of the hubs we have operate on
a one node per user model - a user gets all the resources on a
single node. The only exceptions to this are the pangeo-hubs
(which we should investigate) and utoronto. So this PR
disables user scheduler by default, but enables it on just those
two hubs.

We should document *when* the user scheduler should be explicitly
turned on.