Stop passing a JupyterHub API token for use by health check
Recently the deployer script was made to acquire a JupyterHub API token
by inspecting the `hub` k8s Secret created by the JupyterHub Helm chart
for when it was to interact with the JupyterHub API to make health
checks.

As this API token in modern versions of the JupyterHub Helm chart will
be generated automatically for `hub.services` registered JupyterHub
services, or reused in case it already is defined in the k8s Secret, we
are able to stop defining this token explicitly over and over.

In other words:

- For new installations: the API token will be generated by the
  JupyterHub Helm chart instead
- For existing installations: the API token already set in the k8s
  Secret `hub` will be re-used by helm 3's `lookup` function as used by
  the JupyterHub Helm chart.