Setup dex to proivde oidc authentication
JupyterHub provides oauth2, but *not* oidc. This sucks, but
as with most problems in computer science, we can solve this
with another layer of indirection.

In this case, we use https://dexidp.io to setup an oidc provider
that is itself authenticated via JupyterHub. This is transparent
to the user (because we don't require any consent screens), so in
practice it's gonna look like JupyterHub is providing oidc itself.

Instead of automatically injecting config from the deployer,
we're instead manually copy-pasting config into the .values
files directly. This helps with right to replicate, and keeps our
deployer smaller