Add support for pulling from private github repos
Instead of using deploy keys, we use
https://github.com/yuvipanda/git-credential-helpers,
which is also used for pulling from private repos with nbgitpuller.

I've created a new GitHub app for the 2i2c staging hub,
at https://github.com/apps/2i2c-staging-hub-static-pull and 'installed'
it for the private repo at https://github.com/yuvipanda/test-repo-push.
This allows the static-sites pod to pull it, and it's available at
https://staging.2i2c.cloud/textbook/ now