Tweak dask-gateway on LEAP hub
- Provide more memory & cpu for the dask-scheduler pod. When
  we were running a larger cluster, the scheduler kept dying
  and providing strange errors. Increasing resources helps us
  get bigger clusters.
- Tweak the size of nodes for dask. Dask, at least as used in
  LEAP, is primarily memory bound and not cpu-bound. This helps
  us use resources more efficiently.

Worked with @TomNicholas and @rabernat for bringing up a big cluster
to crunch through some compute for Tom's talk today.

The other thing we discovered is that by default, dask-gateway
sets number of *threads* to number of CPUs, which actually causes
a lot of GIL contention in our workflows. Need to carefully tune
and investigate that.