Change how `comment` variable is defined
In CI/CD, this was breaking due to `comment` not being set, which
I don't believe should happen using the next(...) function. This
PR makes the code look more like the block that extracts the
artifact ID further up in the code in the hopes that this will
alleviate this issue.