Keep a placeholder node around for UToronto Hub
Users were reporting slow starts and timeouts in the morning,
as many users come on at the same time. This change keeps
a placeholder *node* around, with a placeholder pod that
will get displaced whenever a user needs that node. This should
increase the odds of a new node being up by the time more
than 1 node worth of users pop in.

Ref https://2i2c.freshdesk.com/a/tickets/201