Allow restricting Callysto logins by domain
With Callysto, we do not use email addresses to try preserve
user privacy a little. However, this makes restricting the domains
users can login be a bit difficult - all our validation code
validates the username itself, rather than any other attribute of
it.

Here, we customize the authenticator to validate the email of the
user, but not use that as the username. This allows us to enforce
a list of domains (including wildcards) that can be allowed to login
without having to set their emails to be their usernames.

Ref https://github.com/2i2c-org/infrastructure/issues/1678