Update image and adjust admin list
Some of our notebooks use the mobilechelonian module which seems to have
broken with the release of ipywidgets 8. I've pinned ipywidgets at 7.x
in our image while we investigate and this change uses that image for
our hub. At the same time I've adjusted the list of admin users to match
the Callysto admin team from Cybera and PIMS.

Signed-off-by: Ian Allison <iana@pims.math.ca>