Give all students equal resources during exam
During an exam, it is important that all students have access
to the same computational resources. 2G should be enough, given
it was the limit for everyone in general. We are *adding* a CPU
limit here, where we did not have one before - so this must
be validated beforehand to make sure it doesn't have adverse
effects.

Ref https://github.com/2i2c-org/infrastructure/issues/1905