Add deployment to measure free space in hub home dir
This adds a deployment with a single pod of
[node_exporter](https://github.com/prometheus/node_exporter) running,
configured only to report filesystem statistics. We mount the same
home directory configuration that is used in the hub to this pod,
so it will provide us with information on total disk size and
used disk. This would allow us to have a Grafana panel with
'disk used %' for each of these hubs! We can even alert on those
later.