Add two more redirects for UToronto
In https://2i2c.freshdesk.com/a/tickets/331, UToronto
is going to be using *.datatools.utoronto.ca as the domain
for most hubs. The current hub is at jupyter.utoronto.ca,
and will eventually be at jupyter.datatools.utoronto.ca. This
sets up a redirect from that *future* domain to current domain
so the future domain can also be advertised now.