Stop forcing use of NotebookApp
This was a temporary fix as a PR landed in JupyterHub &
we tried to upgrade JupyterHub in-image everywhere. As this
just needs to be JupyterHub 2.3.1 to fix the bug we were guarding
against, I think we're safe to get rid of this now.

The specific issue we were guarding against was that selecting
'rstudio' as the default URL will not work - or in general,
setting the default URL won't work, and everyone will just land
in JupyterLab. If we can keep an eye out for that, we're good.
If we do run into that, we can both upgrade JupyterHub package
in their image, or bring back this env var temporarily.

Fixes https://github.com/2i2c-org/infrastructure/issues/2156