Update LEAP hub images to latest pangeo image
This brings in a new dask-gateway version to match what
we have in the chart, which might help with some of the
issues the LEAP users have been facing.

This also brings in a newer version of JupyterLab, which
requires we stop using NotebookApp. This overrides it for
this specific hub until
https://github.com/2i2c-org/infrastructure/pull/2160
is merged.

Ref https://2i2c.freshdesk.com/a/tickets/433