terraform: relax pins of providers by pinning to major versions only
By pinning to the patch or minor versions, we may miss out on bugfixes
and additions of relevance.

In this commit, I've made use of the `~>` version specifier, which means
that the last provided version specifier is a lower bound. So, `~> 3.0`
means `>= 3.0, < 4.0` and `~> 0.7.2` means `>= 0.7.2, < 0.8`.

I see this as a good trade off between pinning tightly and not pinning
at all. Like this we can avoid bugs, rely on up-to-date documentation,
without checking when something is introduced, and avoid breaking changes
quite well.

In this commit, I've also updated the lower bound to the most recent
version to help us narrow the range of actively used versions.