leap: stop pulling the tensorflow image (slows down other startups)
This was an optimization for a workshop where I assume nodes were
pre-started and made sense for that, but would slow down startup on new
nodes if not this specific image was requested on a non-pre-warmed node
after the workshop.

Since the workshop has ended, we can safely remove this now.

Related tickets:
- Workshop: https://2i2c.freshdesk.com/a/tickets/349
- Slow startup discussed: https://2i2c.freshdesk.com/a/tickets/435