Switch utoronto R hub to use CILogon
AzureAD has given us trouble in the past too. This makes that
CILogon's problem, and the current issues with AzureAD don't
seem to happen here. This can't currently work for the main
hub because we use opaque oid as user identifier, rather than
email. If we can make a mapping, that should also be possible.

Ref https://github.com/2i2c-org/infrastructure/issues/2628