basehub, jupyterhub: remove outdated and redundant network-access
The prometheus-server's pod labels changed at some point, and since this
reule related to those labels this rule stopped working. That must have
been when we started having failures to see this data.

Now though, we shouldn't need this rule as we can rely on dedicated
network-access labels respected by the hub's NetworkPolicy.