Deploy GPU nodes across more zones
In https://2i2c.freshdesk.com/a/tickets/764, LEAP users are
running out of GPU on the one zone their notebook nodes are in.
This just expands that to all possible zones just for GPU nodes,
to maximize the amount of GPUs made available to them. This comes
at the cost of home directory access maybe being slightly slower,
but that's ok.