Switch to jupyter/scipy-notebook as 'default'
https://github.com/2i2c-org/infrastructure/issues/2336 lists
issues with the current 'default' image, and the history of
how it came to be. The primary reason for its existence was
addressed in https://github.com/2i2c-org/infrastructure/issues/2435,
so it fundamentally does not really need to exist anymore.

However, there are still 4 hubs using that image, inherited
from the default. This PR does the following:

- Switches to jupyter/scipy-notebook as the 'default', when
  nothing else is specified. Note that because all our hubs
  have an image specified *somewhere* except for these 4
  (see the spreadsheet linked to from https://github.com/2i2c-org/infrastructure/issues/2582),
  this will actually have *no effect* for existing hubs at
  all! Just means that *future* hubs will get this image as
  the default.
- Explicitly sets the image for the 4 hubs still using the
  old 'default' image. These communities will need to be
  reached out to, and the image changed.

After this, I believe we can archive the old 'default' image!

Ref https://github.com/2i2c-org/infrastructure/issues/2336