Stop prometheus from balooning memory usage on startup
https://github.com/prometheus/prometheus/issues/6934 is
the core problem, and we can afford to lose a tiny bit of
data during restarts by killing the wal at startup instead of
giving prometheus a lot of memory to read it all in. We
lose data during the WAL reading phase anyway, and the amount
of memory required seems ginormous (the linked issue has folks
topping out at 64G). Apparently 2.45 is supposed to help,
but it's not out yet (see
https://github.com/prometheus/prometheus/issues/6934#issuecomment-1536172431).

We should only do the wal deletions on prometheii that have this
specific problem.

I've removed the higher memory request for this, and brought it
in line with our other prometheii - and it seems to work fine
without any issues. I've thus also reduced the size of the
core node, practically reverting https://github.com/2i2c-org/infrastructure/pull/2701