Add gp3 storageclass
We add this unconditionally to all clusters for simplification,
so we can set storageClass: gp3 for new clusters that come up on
AWS without issue. This doesn't change the default, and does not
change the storageclass in existing clusters. In addition to using
gp3, it also sets reclaimPolicy to Retain, so if the PVC is deleted,
it does not delete the PV or the underlying EBS volume.

Ref https://github.com/2i2c-org/infrastructure/issues/2906
Ref https://github.com/2i2c-org/infrastructure/issues/2717