Move CosmicDS to AWS hub
The dashboards are a affected at first load by the network slowdown
being debugged in
https://github.com/2i2c-org/infrastructure/issues/2947#issuecomment-1685985572.
Originally we thought it only affects uploads, but definitely seems
to affect downloads too.

That seems to be actually the *only* cluster affected, so we can
temporarily move them to the AWS cluster to see how that works.
There is no persistent storage nor user management, so this is
a simple move.

Ref https://github.com/2i2c-org/infrastructure/issues/2986