Enable using cosmicds jupyterhub as an authentication *provider*
OAuth2 credentials for the CosmicDS portal, which uses
this JupyterHub as authentication *provider*. So when users hit
"Login" in the CosmicDS portal, they're actually redirected to this
JupyterHub (via auth0). This ensures that the portal knows exactly
the (anonymized) usernames of these users, and can do additional work
on their part to track them as necessary.

From perspective of 2i2c engineers, this is just another JupyterHub
service - just one that we don't have to maintain or manage! Our
responsibility is to *provide the credentials*, and no more.

Documentation of this feature is coming in a separate PR,
so as to not block this one from getting merged.

Ref https://github.com/2i2c-org/infrastructure/issues/3019