Alert on GCP when home directory has < 512GB
This weekend too we had an outage caused by a full disk.
In the past we have gotten lucky and caught it before it happened.

Here, we instead alert whenever the filestore has <512GB in it.
This alert will be sent to a pagerduty channel I've configured
to post on our slack.

I've applied this already to the shared cluster, and when
merged, I'll apply this to all our GCP clusters.

I also checked the pricing - this shouldn't cost us anything,
as it's an alert policy on a pre-existing metric.

Fixes https://github.com/2i2c-org/infrastructure/issues/3084