Define repo root and helm chart dirs in one place than use that directly
Define two most used file locations (the root of the repo and the
location of helm charts) inside the file_aquisition.py file and
import and use them in all the places there were initially redifined.

This way, we will need to update them only once and not everytime a file
that uses them changes location.