Grant access for 2i2c members only via admin_users
With oauthenticator 16, being listed in `admin_users` makes 2i2c members
have access even if they are now listed in `allowed_domains` or
`allowed_organizations`. Due to this, we can avoid the need to declare
`2i2c.org` and `2i2c-org` there and rely on being listed in
`admin_users`.