Allow JupyterLab to show hidden files
This doesn't change any behavior, but just allows JupyterLab
to show a 'View -> Show Hidden Files' option that folks can
turn on. The only reason it isn't on by default is apparently
a historical
accident (https://github.com/jupyterlab/jupyterlab/issues/11304#issuecomment-945466766)

Instead of turning it on piecemeal per-hub, this just turns
it on for everyone. There's no reason to not.

Ref https://2i2c.freshdesk.com/a/tickets/1056