Allow public access to buckets via IAM, not ACL
Previously, we were mixing ACL and IAM, which led to
basically the bucket *not* being accessible publicly - only to
authenticated users.

This switches everything to using IAM, which *does* make the
bucket properly publicly accessible.

In addition, there's now a policy that we only enable this
when 2i2c is *not* handling billing, as there can be
disastrous cost consequences.

We fix this for the LEAP bucket.

Config is moved to `user_buckets` rather than `hub_permissions`,
as the config is purely set on the bucket and not related to
which hub we are configuring.

Ref https://2i2c.freshdesk.com/a/tickets/954