LEAP: Enable usage logs for GCP buckets
In https://2i2c.freshdesk.com/a/tickets/1105, we were
requested to turn on [usage
logs](https://cloud.google.com/storage/docs/access-logs)
for the GCP buckets in the LEAP project. This moves that
config into terraform, but also makes it clearer that while
we can enable this config, we currently can not actually help
folks *parse* the logs. But the data format is simple
enough that they should be able to.