cloudbank: Explicitly specify what nodepools users should go to
Otherwise, they still end up on the old 'user' nodepool, and it
can not be decomissioned. With this, they will end up on the
equivalent 'new' nodepool, and then I can get rid of the old one.

I have also had to adjust the CPU limit, as at some point I suppose
we switched from 8 CPU nodes to 4 but did not adjust this.