Redirect all UToronto home pages to unified homepage
UToronto will have a new URL (to be determined...) that they
will manage, and will act as the 'unified' home page. In
https://2i2c.freshdesk.com/a/tickets/952 they have asked us
to redirect anyone landing just on the home page of the hubs
to this new URL. I've setup the [utoronto
branch](https://github.com/2i2c-org/default-hub-homepage/tree/utoronto)
of the home pages repo to redirect to the new URL.

This just deploys that change. The new URL hasn't been provided
to us yet, so the utoronto branch in the default-hub-homepage
repo will need to be updated once that URL is present.

Ref https://github.com/2i2c-org/infrastructure/issues/2729