Check for technical support contacts as part of new hub turn-up
As part of https://github.com/2i2c-org/infrastructure/issues/3048#issuecomment-1907195217,
we now have an airtable be the 'source of truth' for who is
allowed to initiate support requests. This PR adds an item in
the checklist for new hub turn up, making sure we validate that
there *is* a technical contact for this hub.

The airtable in use may change soon as well as the terminology
(as partnerships works on it https://github.com/2i2c-org/infrastructure/issues/3048#issuecomment-1909011115).
But it is important to have a verification step here, so we don't
let whatever airtable it is be out of date.

Ref https://github.com/2i2c-org/infrastructure/issues/3048#issuecomment-1907335901