Add runbook for setting up a community managed hub domain name
Documenting in detail the process being followed in
https://2i2c.freshdesk.com/a/tickets/1280. The audience is anyone
handling support.

There already was a CNAME management howto - however, it was not
structured as a *runbook* that someone could just follow linearly.
The new guide is structured better, has all the information from
the previous guide and some more.

In addition, it also codifies Erik's request in
https://github.com/2i2c-org/infrastructure/issues/2356 to always
set up redirects from a 2i2c managed domain name to the community
managed domain.

The section on changing authorized URL in CILogon is missing because I don't
know how to do that - but I think that's ok! We can add that the
next time we change it. These guides don't have to be 100%
complete to be useful.

https://github.com/2i2c-org/infrastructure/pull/3680 is the
associated PR implementing this. Note that we are debugging some
auth0 related things there with the community - I'll update this
guide once there's more clarity there, but that should not block
this PR.