Fix extra case sensitivity in GitHub teams check
The community reproted an issue with people from the new organization
being unable to login. This was the error message, only found in the logs.

> Your GitHub team membership is insufficient to launch any server profiles.
>
> GitHub teams you are a member of that this JupyterHub knows about are
> nasa-openscapes-workshops:workshopaccess-2i2c, <others>
>
> If you are part of additional teams, log out of this JupyterHub and
> log back in to refresh that information

Looking at the *case* of the error message, we see that
our config specifies `nasa-openscapes-workshops:WorkshopAccess-2i2c`
while the team name returned to us by GitHub is all lower case instead.

GitHub orgs are [not case sensitive](https://docs.github.com/en/rest/orgs/orgs?apiVersion=2022-11-28#get-an-organization),
and it was not clear to me if *teams* are or are not. Nevertheless,
I verified that you can not actually create a different team in
the same org with the same name but different case.

This patch casefolds all the team and org names we use, so
comparisons are case insensitive. casefold works for all languages,
while calling `.lower()` only works for a subset of languages.

Ref https://github.com/2i2c-org/infrastructure/issues/3695